Renamed to WFCollabProject, changed placements of code
