+++ WFProject/WFCollabProject/bin/Debug/net7.0-windows/WFCollabProject.runtimeconfig.json
@@ -1,0 +1,15 @@
+{
+  "runtimeOptions": {
+    "tfm": "net7.0",
+    "frameworks": [
+      {
+        "name": "Microsoft.NETCore.App",
+        "version": "7.0.0"
+      },
+      {
+        "name": "Microsoft.WindowsDesktop.App",
+        "version": "7.0.0"
+      }
+    ]
+  }
+}
